requirements.txt
Flask-SocketIO
python-dotenv
openai
